ai/orchestrator.py
# AI Orchestration Script
# Placeholder for Vryndara AI Kernel

import os
from langchain.llms import OpenAI
from crewai import Agent, Task, Crew

# Load environment variables
from dotenv import load_dotenv
load_dotenv()

llm = OpenAI(temperature=0.7)

# Define agents
analyst = Agent(
    role="Data Analyst",
    goal="Analyze IoT sensor data",
    backstory="Expert in processing telemetry data",
    llm=llm
)

executor = Agent(
    role="Infrastructure Executor",
    goal="Execute autonomous commands",
    backstory="Handles physical infrastructure control",
    llm=llm
)

# Define tasks
analyze_task = Task(
    description="Analyze incoming sensor data for anomalies",
    agent=analyst
)

execute_task = Task(
    description="Execute appropriate actions based on analysis",
    agent=executor
)

# Create crew
crew = Crew(agents=[analyst, executor], tasks=[analyze_task, execute_task])

if __name__ == "__main__":
    result = crew.kickoff()
    print(result)