ai/requirements.txt
langchain==0.1.0
crewai==0.1.0
openai==1.3.0
python-dotenv==1.0.0